+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+- Adds [`Oolong.isObject`][isObject].
+
+[isObject]: https://github.com/moll/js-oolong/blob/master/doc/API.md#Oolong.isObject
+
 ## 1.9.0 (Jun 1, 2015)
 - Adds [`Oolong.wrap`][wrap].